backend/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
pymongo>=4.6.1
sentence-transformers>=2.2.2
google-generativeai>=0.2.0
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
gunicorn>=23.0.0
whitenoise==6.7.0
numpy>=1.24.0
